requirements.txt
numpy
scipy
tensorflow-gpu==1.12
librosa
tqdm